prototype/requirements.txt
flask==3.0.3
google-genai==1.0.0
gunicorn==22.0.0
python-dotenv==1.0.1
